refactor(frontend): move form submission into TanStack-backed hooks
useEventDraftSubmit / useEventEditSubmit assemble the payload from the
form atoms, run the create/update mutation, and navigate on success,
leaving the form components thin (same pattern as useEventListSearch).
Also document why the scoped Provider / hydrate split into a separate
content component is required.
